Refactor the importers
Refactor the importers so that the duplication of the codes can
be removed.

Refs HP-61
